backend/requirements.txt
# backend\requirements.txt

aiogram>=3.13.0
fastapi>=0.115.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
asyncpg>=0.29.0
redis>=5.0.0
SQLAlchemy>=2.0.29
alembic>=1.13.0
aiohttp>=3.9.0
aioboto3>=13.1.0